ml/requirements.txt
langchain-community 
tavily-python
youtube_search
python-dotenv
fastapi[standard]
uvicorn
openai
pillow
duckduckgo_search
sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
llama-index
sentence-transformers
langchain
fastembed 
tiktoken
python-dotenv
pandas
ipykernel
setuptools


